feat: cleanup and use feeBps from confirmedQuote
