+++ src/pages/ThorChainLP/components/ReusableLpStatus/TransactionRow.tsx
@@ -22,8 +22,6 @@ import type { SendInput } from 'components/Modals/Send/Form'
 import { estimateFees, handleSend } from 'components/Modals/Send/utils'
 import { Row } from 'components/Row/Row'
 import { useWallet } from 'hooks/useWallet/useWallet'
-import { bn } from 'lib/bignumber/bignumber'
-import { calculateFees } from 'lib/fees/model'
 import { toBaseUnit } from 'lib/math'
 import { assetIdToPoolAssetId } from 'lib/swapper/swappers/ThorchainSwapper/utils/poolAssetHelpers/poolAssetHelpers'
 import { assertUnreachable, isToken } from 'lib/utils'
@@ -39,7 +37,6 @@ import { THORCHAIN_POOL_MODULE_ADDRESS } from 'lib/utils/thorchain/constants'
 import type { AsymSide, ConfirmedQuote } from 'lib/utils/thorchain/lp/types'
 import { depositWithExpiry } from 'lib/utils/thorchain/routerCalldata'
 import { getThorchainLpPosition } from 'pages/ThorChainLP/queries/queries'
-import { selectIsSnapshotApiQueriesPending, selectVotingPower } from 'state/apis/snapshot/selectors'
 import {
   selectAccountNumberByAccountId,
   selectAssetById,
@@ -75,12 +72,6 @@ export const TransactionRow: React.FC<TransactionRowProps> = ({
   const queryClient = useQueryClient()
   const translate = useTranslate()
   const selectedCurrency = useAppSelector(selectSelectedCurrency)
-  const votingPower = useAppSelector(selectVotingPower)
-  const isSnapshotApiQueriesPending = useAppSelector(selectIsSnapshotApiQueriesPending)
-  const isVotingPowerLoading = useMemo(
-    () => isSnapshotApiQueriesPending && votingPower === undefined,
-    [isSnapshotApiQueriesPending, votingPower],
-  )
   // TOOO(gomes): we may be able to handle this better, or not
   const asset = useAppSelector(state => selectAssetById(state, assetId ?? ''))
   const poolAsset = useAppSelector(state => selectAssetById(state, poolAssetId ?? ''))
@@ -243,11 +234,6 @@ export const TransactionRow: React.FC<TransactionRowProps> = ({
   const handleSignTx = useCallback(() => {
     const isRuneTx = asset?.assetId === thorchainAssetId
 
-    const { feeBps } = calculateFees({
-      tradeAmountUsd: bn(confirmedQuote.totalAmountFiat),
-      foxHeld: votingPower !== undefined ? bn(votingPower) : undefined,
-    })
-
     if (
       !(
         assetId &&
@@ -291,9 +277,9 @@ export const TransactionRow: React.FC<TransactionRowProps> = ({
             if (runeAccountNumber === undefined) throw new Error(`No account number found for RUNE`)
 
             const adapter = assertGetThorchainChainAdapter()
-            const memo = `+:${thorchainNotationAssetId}:${
-              otherAssetAddress ?? ''
-            }:ss:${feeBps.toFixed(0)}`
+            const memo = `+:${thorchainNotationAssetId}:${otherAssetAddress ?? ''}:ss:${
+              confirmedQuote.feeBps
+            }`
 
             const estimatedFees = await estimateFees({
               cryptoAmount: amountCryptoBaseUnit,
@@ -342,9 +328,9 @@ export const TransactionRow: React.FC<TransactionRowProps> = ({
                 ? getAddress(fromAssetId(assetId).assetReference)
                 : '0x0000000000000000000000000000000000000000'
 
-              const memo = `+:${thorchainNotationAssetId}:${
-                otherAssetAddress ?? ''
-              }:ss:${feeBps.toFixed(0)}`
+              const memo = `+:${thorchainNotationAssetId}:${otherAssetAddress ?? ''}:ss:${
+                confirmedQuote.feeBps
+              }`
               const amount = BigInt(amountCryptoBaseUnit.toString())
 
               return { memo, amount, expiry, vault, asset }
@@ -371,7 +357,6 @@ export const TransactionRow: React.FC<TransactionRowProps> = ({
               wallet,
             })
 
-            // TODO(gomes): fees estimation
             const txid = await buildAndBroadcast({
               adapter,
               buildCustomTxInput,
@@ -386,9 +371,9 @@ export const TransactionRow: React.FC<TransactionRowProps> = ({
             // ATOM/RUNE/ UTXOs- obviously make me a switch case for things to be cleaner
             if (!accountAssetAddress) throw new Error('No accountAddress found')
 
-            const memo = `+:${thorchainNotationAssetId}:${
-              otherAssetAddress ?? ''
-            }:ss:${feeBps.toFixed(0)}`
+            const memo = `+:${thorchainNotationAssetId}:${otherAssetAddress ?? ''}:ss:${
+              confirmedQuote.feeBps
+            }`
 
             const estimateFeesArgs = {
               cryptoAmount: amountCryptoPrecision,
@@ -435,8 +420,6 @@ export const TransactionRow: React.FC<TransactionRowProps> = ({
     })
   }, [
     asset,
-    confirmedQuote.totalAmountFiat,
-    votingPower,
     assetId,
     poolAssetId,
     poolAsset,
@@ -447,6 +430,7 @@ export const TransactionRow: React.FC<TransactionRowProps> = ({
     amountCryptoPrecision,
     runeAccountNumber,
     otherAssetAddress,
+    confirmedQuote.feeBps,
     assetAccountNumber,
     accountAssetAddress,
     selectedCurrency,
@@ -497,9 +481,7 @@ export const TransactionRow: React.FC<TransactionRowProps> = ({
             size='lg'
             colorScheme='blue'
             onClick={handleSignTx}
-            isLoading={
-              status === TxStatus.Pending || isInboundAddressLoading || isVotingPowerLoading
-            }
+            isLoading={status === TxStatus.Pending || isInboundAddressLoading}
           >
             {translate('common.signTransaction')}
           </Button>
